Lookup table switch handler

A function to call an epidm lookup table and recode where we are aware of a new value.

Built in are the organism re-classifications and specimen_type groupings and a manual mode.

Usage

lookup_recode(
  src,
  type = c("species", "specimen", "inpatient_admission_method",
    "inpatient_discharge_destination", "ecds_destination_code", "manual"),
  .import = NULL
)

Arguments

src

a character, vector or column containing the value(s) to be referenced

type

a character value to denote the lookup table used

.import

a list in the order list(new,old) containing the values for another lookup table existing in the environment

Value

a list object of the recoded field

Examples

df <- data.frame(
  spec = c(
    sample(grep(")",
                respeciate_organism$previous_organism_name,
                value=TRUE,
                invert = TRUE),
           9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  type = sample(specimen_type_grouping$specimen_type,12),
  date = sample(seq.Date(from = Sys.Date()-365,
                         to = Sys.Date(),
                         by = "day"),12)
)
df <- df[order(df$date),]

# show the data before the changes
df
#>                              spec                 type       date
#> 6        PSEUDOMONAS PAUCIMOBILIS               RECTUM 2023-08-26
#> 1          ACTINOBACULUM SCHAALII       SYNOVIAL FLUID 2023-10-10
#> 12                 CANDIDA AUREUS    TRACHEAL ASPIRATE 2023-11-01
#> 9    TETRATHIOBACTER KASHMIRENSIS             DUODENUM 2024-01-09
#> 2           STREPTOCOCCUS GROUP D            BRONCHIAL 2024-01-10
#> 7       CHRYSEOBACTERIUM MIRICOLA             PERIANAL 2024-01-13
#> 4         EUBACTERIUM AEROFACIENS  LOWER GENITAL TRACT 2024-01-22
#> 3         PROPIONIBACTERIUM ACNES   MIDDLE EAR/MASTOID 2024-01-27
#> 10               ESCHERICHIA COLI PUS (SOURCE UNKNOWN) 2024-03-12
#> 8  LEPTOSPIRA ICTEROHAEMORRHAGIAE                  EMU 2024-04-21
#> 5          ALCALIGENES PIECHAUDII           BIOPSY-NOS 2024-06-17
#> 11                     SARS-COV-2               SPLEEN 2024-07-05

# check the lookup tables
# observe the changes
head(respeciate_organism[1:2])
#>                  previous_organism_name       organism_species_name
#> 1             ALCALIGENES DENITRIFICANS ACHROMOBACTER DENITRIFICANS
#> 2                ALCALIGENES PIECHAUDII    ACHROMOBACTER PIECHAUDII
#> 3              ALCALIGENES XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 4 ALCALIGENES XYLOSOXIDANS XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 5                ACTINOBACULUM SCHAALII       ACTINOTIGNUM SCHAALII
#> 6          TETRATHIOBACTER KASHMIRENSIS      ADVENELLA KASHMIRENSIS
df$species <- lookup_recode(df$spec,'species')
df[,c('spec','species','date')]
#>                              spec                   species       date
#> 6        PSEUDOMONAS PAUCIMOBILIS SPHINGOMONAS PAUCIMOBILIS 2023-08-26
#> 1          ACTINOBACULUM SCHAALII     ACTINOTIGNUM SCHAALII 2023-10-10
#> 12                 CANDIDA AUREUS            CANDIDA AUREUS 2023-11-01
#> 9    TETRATHIOBACTER KASHMIRENSIS    ADVENELLA KASHMIRENSIS 2024-01-09
#> 2           STREPTOCOCCUS GROUP D     ENTEROCOCCUS FAECALIS 2024-01-10
#> 7       CHRYSEOBACTERIUM MIRICOLA  ELIZABETHKINGIA MIRICOLA 2024-01-13
#> 4         EUBACTERIUM AEROFACIENS   COLLINSELLA AEROFACIENS 2024-01-22
#> 3         PROPIONIBACTERIUM ACNES       CUTIBACTERIUM ACNES 2024-01-27
#> 10               ESCHERICHIA COLI          ESCHERICHIA COLI 2024-03-12
#> 8  LEPTOSPIRA ICTEROHAEMORRHAGIAE    LEPTOSPIRA INTERROGANS 2024-04-21
#> 5          ALCALIGENES PIECHAUDII  ACHROMOBACTER PIECHAUDII 2024-06-17
#> 11                     SARS-COV-2                SARS-COV-2 2024-07-05

head(specimen_type_grouping)
#>      specimen_type specimen_group
#> 1            BLOOD          Blood
#> 2       CORD BLOOD          Blood
#> 3            JOINT Bones & Joints
#> 4   SYNOVIAL FLUID Bones & Joints
#> 5             BONE Bones & Joints
#> 6 JOINT PROSTHESIS Bones & Joints
df$grp <- lookup_recode(df$type,'specimen')
df[,c('species','type','grp','date')]
#>                      species                 type                     grp
#> 6  SPHINGOMONAS PAUCIMOBILIS               RECTUM       Faeces & Lowergut
#> 1      ACTINOTIGNUM SCHAALII       SYNOVIAL FLUID          Bones & Joints
#> 12            CANDIDA AUREUS    TRACHEAL ASPIRATE           URT/Mouth/Ear
#> 9     ADVENELLA KASHMIRENSIS             DUODENUM       Faeces & Lowergut
#> 2      ENTEROCOCCUS FAECALIS            BRONCHIAL Lower Respiratory Tract
#> 7   ELIZABETHKINGIA MIRICOLA             PERIANAL       Faeces & Lowergut
#> 4    COLLINSELLA AEROFACIENS  LOWER GENITAL TRACT                 Genital
#> 3        CUTIBACTERIUM ACNES   MIDDLE EAR/MASTOID           URT/Mouth/Ear
#> 10          ESCHERICHIA COLI PUS (SOURCE UNKNOWN)                  Fluids
#> 8     LEPTOSPIRA INTERROGANS                  EMU            Urine/Kidney
#> 5   ACHROMOBACTER PIECHAUDII           BIOPSY-NOS           Swabs-General
#> 11                SARS-COV-2               SPLEEN                 Tissues
#>          date
#> 6  2023-08-26
#> 1  2023-10-10
#> 12 2023-11-01
#> 9  2024-01-09
#> 2  2024-01-10
#> 7  2024-01-13
#> 4  2024-01-22
#> 3  2024-01-27
#> 10 2024-03-12
#> 8  2024-04-21
#> 5  2024-06-17
#> 11 2024-07-05

# for a tidyverse use
# df %>% mutate(spec=lookup_recode(spec,'species))

# manual input of your own lookup
# .import=list(new,old)
lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
#> [1] "ACHROMOBACTER DENITRIFICANS"