Respeciate unspecified samples

Some samples within SGSS are submitted by laboratories as "GENUS SP" or "GENUS UNNAMED". However, they may also have a fully identified sample taken from the same site within a recent time period. This function captures species_col from another sample within X-days of an unspeciated isolate.

Usage

respeciate_generic(
  x,
  group_vars,
  species_col,
  date_col,
  window = c(0:Inf),
  .forceCopy = FALSE
)

Arguments

x

a data.frame or data.table object

group_vars

the minimum grouping set of variables for like samples in a character vector; suggest c('patient_id','specimen_type','genus')

species_col

a character containing the column with the organism species_col name

date_col

a character containing the column with the specimen/sample date_col

window

an integer representing the number of days for which you will allow a sample to be respeciated

.forceCopy

default FALSE; TRUE will force data.table to take a copy instead of editing the data without reference

Value

a data.table with a recharacterised species_col column

Examples

df <- data.frame(
ptid = c(round(runif(25,1,5))),
spec = sample(c("KLEBSIELLA SP",
                "KLEBSIELLA UNNAMED",
                "KLEBSIELLA PNEUMONIAE",
                "KLEBEIELLA OXYTOCA"),
              25,replace = TRUE),
type = "BLOOD",
specdate = sample(seq.Date(Sys.Date()-21,Sys.Date(),"day"),25,replace = TRUE)
)

respeciate_generic(x=df,
                   group_vars=c('ptid','type'),
                   species_col='spec',
                   date_col='specdate',
                   window = 14)[]
#> 14-day round 1: 7 SP or UNNAMMED isolates respeciated
#> 14-day round 2: 3 SP or UNNAMMED isolates respeciated
#> 14-day round 3: 3 SP or UNNAMMED isolates respeciated
#>      ptid                  spec   type   specdate
#>     <num>                <char> <char>     <Date>
#>  1:     1    KLEBSIELLA UNNAMED  BLOOD 2024-07-17
#>  2:     1    KLEBSIELLA UNNAMED  BLOOD 2024-07-21
#>  3:     1    KLEBSIELLA UNNAMED  BLOOD 2024-07-21
#>  4:     2    KLEBEIELLA OXYTOCA  BLOOD 2024-07-12
#>  5:     2    KLEBEIELLA OXYTOCA  BLOOD 2024-07-15
#>  6:     2    KLEBEIELLA OXYTOCA  BLOOD 2024-07-22
#>  7:     2    KLEBEIELLA OXYTOCA  BLOOD 2024-07-23
#>  8:     2    KLEBEIELLA OXYTOCA  BLOOD 2024-07-23
#>  9:     2    KLEBSIELLA UNNAMED  BLOOD 2024-07-28
#> 10:     3    KLEBSIELLA UNNAMED  BLOOD 2024-07-11
#> 11:     3 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-12
#> 12:     3 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-14
#> 13:     3 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-16
#> 14:     3 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-18
#> 15:     3    KLEBEIELLA OXYTOCA  BLOOD 2024-07-19
#> 16:     3    KLEBEIELLA OXYTOCA  BLOOD 2024-07-19
#> 17:     3    KLEBEIELLA OXYTOCA  BLOOD 2024-07-23
#> 18:     3    KLEBEIELLA OXYTOCA  BLOOD 2024-07-29
#> 19:     3 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-31
#> 20:     4    KLEBEIELLA OXYTOCA  BLOOD 2024-07-19
#> 21:     4    KLEBEIELLA OXYTOCA  BLOOD 2024-07-28
#> 22:     4    KLEBEIELLA OXYTOCA  BLOOD 2024-07-29
#> 23:     4    KLEBEIELLA OXYTOCA  BLOOD 2024-07-30
#> 24:     5 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-13
#> 25:     5 KLEBSIELLA PNEUMONIAE  BLOOD 2024-07-14
#>      ptid                  spec   type   specdate