Patient ID record grouping

Groups patient records from multiple isolates with a single integer patientID by grouping patient identifiers.

Grouping is based on the following stages:

matching nhs number and date of birth

Hospital number & Date of Birth

NHS number & Hospital Number

NHS number & Name

Hospital number & Name

Sex & Date of Birth & Surname

Sex & Date of Birth & Fuzzy Name

Sex & Year and Month of Birth & Fuzzy Name

Postcode & Name

Name Swaps (when first and last name are the wrong way around)

Identifiers are copied over where they are missing or invalid to the grouped records.

Usage

uk_patient_id(
  data,
  id = list(nhs_number = "nhs_number", hospital_number = "patient_hospital_number",
    date_of_birth = "date_of_birth", sex_mfu = "sex", forename = "forename", surname =
    "surname", postcode = "postcode"),
  .useStages = c(1:11),
  .sortOrder,
  .keepValidNHS = FALSE,
  .forceCopy = FALSE
)

Arguments

data

a data.frame or data.table containing the patient data

id

a named list to provide the column names with identifiers, quoted

nhs_number

the patient NHS number

hospital_number

the patient Hospital numbers also known as the local patient identifier

date_of_birth

the patient date of birth

sex_mfu

the patient sex or gender field as Male/Female/Unknown

forename

the patient forename

surname

the patient surname

postcode

the patient postcode

.useStages

optional, default 1:11; set to 1 if you wish patient ID to be assigned cases with the same DOB and NHS number, set to 2 if you wish patient ID to be assigned to cases with the same hospital number (HOS) and DOB, set to 3 if you wish patient ID to be assigned cases with the same NHS and HOS number, set to 4 if you wish patient ID to be assigned cases with the same NHS number and surname, set to 5 if you wish patient ID to be assigned cases with the same hospital number and surname, set to 6 if you wish patient ID to be assigned cases with the same DOB and surname, set to 7 if you wish patient ID to be assigned cases with the same sex and full name, set to 8 if you wish patient ID to be assigned cases with the same sex, DOB and fuzzy name, set to 9 if you wish patient ID to be assigned cases with the same DOB and fuzzy name, set to 10 if you wish patient ID to be assigned cases with the same name and postcode, set to 11 if you wish patient ID to be assigned cases with the same first name or second name in changing order and date of birth.

.sortOrder

optional; a column as a character to allow a sorting order on the id generation

.keepValidNHS

optional, default FALSE; set TRUE if you wish to retain the column with the NHS checksum result stored as a BOOLEAN

.forceCopy

optional, default FALSE; TRUE will force data.table to take a copy instead of editing the data without reference

Value

A dataframe with one new variable:

id

a unique patient id

valid_nhs

if retained using argument .keepValidNHS=TRUE, a BOOLEAN containing the result of the NHS checksum validation

Examples

uk_patient_id(
 data = head(epidm::lab_data,1000),
 id = list(
   nhs_number = 'nhs_number',
   hospital_number = 'local_patient_identifier',
   date_of_birth = 'patient_birth_date',
   sex_mfu = 'sex',
   forename = 'forename',
   surname = 'surname'
 ),
 .sortOrder = 'specimen_date',
 .forceCopy = TRUE
)[]
#>          id nhs_number local_patient_identifier patient_birth_date    sex
#>       <int>     <char>                   <char>             <Date> <char>
#>    1:     1  852542115                  AY45405         1952-03-31      F
#>    2:     2  636618808                  AN80559         1974-10-16      F
#>    3:     3  676931293                  BE25621         1984-11-12      F
#>    4:     4  150995260                  JY96040         1949-05-13      M
#>    5:     5  656066785                   CK3584         1944-02-03      F
#>   ---                                                                    
#>  996:   996  831527085                  HZ50426         1934-06-18      F
#>  997:   997  470415263                  AX81005         1971-08-06      F
#>  998:   998  581923601                  QZ18729         1964-02-14      M
#>  999:   999  982821706                  HD14018         1969-02-22      F
#> 1000:  1000  541747444                  EI64224         1989-02-14      F
#>         surname forename  organism_species_name specimen_date specimen_type
#>          <char>   <char>                 <char>        <Date>        <fctr>
#>    1:   CHITTUM  DARLENE PSEUDOMONAS AERUGINOSA    2020-01-14  URINE/KIDNEY
#>    2:  AL-GHANI    CALIE  STAPHYLOCOCCUS AUREUS    2020-01-14  URINE/KIDNEY
#>    3:     MYERS    DEBRA  KLEBSIELLA PNEUMONIAE    2020-01-14         BLOOD
#>    4:   BLEVANS    DANTE PSEUDOMONAS AERUGINOSA    2020-01-17         BLOOD
#>    5:   GROOVER  NAJEEMA  KLEBSIELLA PNEUMONIAE    2020-02-01         BLOOD
#>   ---                                                                      
#>  996:     BEALS    JAYLA  STAPHYLOCOCCUS AUREUS    2023-12-24         BLOOD
#>  997:  CASTILLO  YOLANDA  KLEBSIELLA PNEUMONIAE    2023-12-25         BLOOD
#>  998: EL-HALABY    RAJAA  STAPHYLOCOCCUS AUREUS    2023-12-26  URINE/KIDNEY
#>  999:   THURMAN      MAY PSEUDOMONAS AERUGINOSA    2023-12-27  URINE/KIDNEY
#> 1000:  KILSTROM     KALI  KLEBSIELLA PNEUMONIAE    2023-12-29         BLOOD
#>       lab_code local_authority_name local_authority_code
#>         <char>               <char>               <char>
#>    1:  ZJ60076              Enfield            E09000010
#>    2:  VJ49160              Salford            E08000006
#>    3:  ID41899        South Norfolk            E07000149
#>    4:  JH70033           Spelthorne            E07000213
#>    5:  PH46212            Sheffield            E08000019
#>   ---                                                   
#>  996:  WI90375       South Lakeland            E07000031
#>  997:   CW7915           Manchester            E08000003
#>  998:  VJ49160                Wigan            E08000010
#>  999:  YK19355            Redbridge            E09000026
#> 1000:  GF74315       South Somerset            E07000189