Link A&E to Inpatient records

Link together ECDS A&E records to HES/SUS inpatient records on NHS number, Hospital Number and Date of Birth. To note that the inpatient records should already be aggregated into spells at the desired level.

link_ae_inpatient(
  ae_data,
  ae_arrival_date,
  ae_departure_date,
  inp_data,
  inp_inp_spell_start_date,
  inp_spell_id,
  nhs_number = c("nhs_number", "nhs_number"),
  hospital_number = c("local_patient_identifier", "local_patient_identifier"),
  patient_dob = c("date_birth", "date_birth"),
  org_code = c("organisation_code_of_provider", "organisation_code_code_of_provider"),
  .forceCopy = FALSE
)

Arguments

ae_data

the ECDS A&E dataset

ae_arrival_date

the ECDS arrival date

inp_data

the HES/SUS inpatient dataset

inp_spell_id

the HES/SUS spell id

nhs_number

a vector containing the column names for the NHS numbers in the order c('ae','inpatient')

hospital_number

a vector containing the column names for the Hospital numbers in the order c('ae','inpatient')

patient_dob

a vector containing the column names for the date of birth in the order c('ae','inpatient')

org_code

a vector containing the column names for the organisation codes in the order c('ae','inpatient')

.forceCopy

a boolean to control if you want to copy the dataset before linking together

ae_ae_departure_date

the ECDS discharge date

admission_date

a string containing the inpatient (HES/SUS) admission date column name

Value

a patient level linked hospital record

See also