Lookup table switch handler

A function to call an epidm lookup table and recode where we are aware of a new value.

Built in are the organism re-classifications and specimen_type groupings and a manual mode.

lookup_recode(src, type = c("species", "specimen", "manual"), .import = NULL)

Arguments

src

a character, vector or column containing the value(s) to be referenced

type

a character value to denote the lookup table used

.import

a list in the order list(new,old) containing the values for another lookup table existing in the environment

Value

a list object of the recoded field

Examples

df <- data.frame(
  sp = c(
    sample(respeciate_organism$previous_organism_name,9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  ty = sample(specimen_type_grouping$specimen_type,12),
  dt = sample(seq.Date(from = Sys.Date()-365,
                       to = Sys.Date(),
                       by = "day"),12)
)
df <- df[order(df$dt),]

df$species <- lookup_recode(df$sp,'species')
df$grp <- lookup_recode(df$ty,'specimen')
df
#>                                                          sp
#> 8                                  ALCALIGENES XYLOSOXIDANS
#> 9                                     ANAEROCOCCUS PREVOTII
#> 10                                         ESCHERICHIA COLI
#> 7      PSEUDOMONAS PAUCIMOBILIS (SPHINGOMONAS PAUCIMOBILIS)
#> 11                                               SARS-COV-2
#> 5                              TETRATHIOBACTER KASHMIRENSIS
#> 6                                            HANIA PARALVEI
#> 4                                     STREPTOCOCCUS GROUP D
#> 12                                           CANDIDA AUREUS
#> 1                                        EUBACTERIUM LENTUM
#> 2  PSYCHROBACTER PHENYLPYRUVICUS (MORAXELLA PHENYLPYRUVICA)
#> 3                                  FLAVIMONAS ORYZIHABITANS
#>                               ty         dt                       species
#> 8                      DRAINSITE 2021-01-05    ACHROMOBACTER XYLOSOXIDANS
#> 9                         SPLEEN 2021-02-01   PEPTOSTREPTOCOCCUS PREVOTII
#> 10                          SWAB 2021-04-27              ESCHERICHIA COLI
#> 7                          STOOL 2021-07-07     SPHINGOMONAS PAUCIMOBILIS
#> 11                     UMBILICUS 2021-09-12                    SARS-COV-2
#> 5                          BLOOD 2021-09-14        ADVENELLA KASHMIRENSIS
#> 6        UPPER RESPIRATORY TRACT 2021-09-18                  HAVNIA ALVEI
#> 4                      ULCER-NOS 2021-11-07         ENTEROCOCCUS FAECALIS
#> 12                     DIALYSATE 2021-11-25                CANDIDA AUREUS
#> 1                            LIP 2021-11-27             EGGERTHELLA LENTA
#> 2                ALVEOLAR LAVAGE 2021-12-18 PSYCHROBACTER PHENYLPYRUVICUS
#> 3  UPPER GASTRO-INTESTINAL TRACT 2022-01-02     PSEUDOMONAS ORYZIHABITANS
#>                        grp
#> 8                   Fluids
#> 9                  Tissues
#> 10           Swabs-General
#> 7        Faeces & Lowergut
#> 11           Swabs-General
#> 5                    Blood
#> 6            URT/Mouth/Ear
#> 4            Swabs-General
#> 12            Urine/Kidney
#> 1            Swabs-General
#> 2  Lower Respiratory Tract
#> 3                  Tissues

# for a tidyverse use
# df %>% mutate(sp=lookup_recode(sp,'species))

lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
#> [[1]]
#> [1] "ACHROMOBACTER DENITRIFICANS"
#>