Lookup table switch handler

A function to call an epidm lookup table and recode where we are aware of a new value.

Built in are the organism re-classifications and specimen_type groupings and a manual mode.

lookup_recode(src, type = c("species", "specimen", "manual"), .import = NULL)

Arguments

src

a character, vector or column containing the value(s) to be referenced

type

a character value to denote the lookup table used

.import

a list in the order list(new,old) containing the values for another lookup table existing in the environment

Value

a list object of the recoded field

Examples

df <- data.frame(
  spec = c(
    sample(grep(")",
                respeciate_organism$previous_organism_name,
                value=TRUE,
                invert = TRUE),
           9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  type = sample(specimen_type_grouping$specimen_type,12),
  date = sample(seq.Date(from = Sys.Date()-365,
                         to = Sys.Date(),
                         by = "day"),12)
)
df <- df[order(df$date),]

# show the data before the changes
df
#>                            spec                    type       date
#> 3          CHRYSEOMONAS LUTEOLA                  VAGINA 2021-02-27
#> 10             ESCHERICHIA COLI              BIOPSY-NOS 2021-04-12
#> 9         STREPTOCOCCUS GROUP D                  CERVIX 2021-04-16
#> 2  TETRATHIOBACTER KASHMIRENSIS                  SALIVA 2021-05-01
#> 7         ANAEROCOCCUS PREVOTII                  THROAT 2021-05-16
#> 11                   SARS-COV-2        JOINT PROSTHESIS 2021-05-24
#> 5              STOMATOCOCCUS SP ENDOTRACHEAL SECRETIONS 2021-08-04
#> 1       PROPIONIBACTERIUM ACNES                PERINEUM 2021-08-07
#> 6       RHODOCOCCUS BRONCHIALIS   ENDOTRACHEAL ASPIRATE 2021-08-21
#> 12               CANDIDA AUREUS       HEART/HEART VALVE 2021-08-25
#> 8             EGGERTHELLA LENTA                    BURN 2021-08-29
#> 4      PSEUDOMONAS PAUCIMOBILIS        WOUND (SURGICAL) 2022-02-07

# check the lookup tables
# observe the changes
head(respeciate_organism[1:2])
#>                  previous_organism_name       organism_species_name
#> 1             ALCALIGENES DENITRIFICANS ACHROMOBACTER DENITRIFICANS
#> 2                ALCALIGENES PIECHAUDII    ACHROMOBACTER PIECHAUDII
#> 3              ALCALIGENES XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 4 ALCALIGENES XYLOSOXIDANS XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 5                ACTINOBACULUM SCHAALII       ACTINOTIGNUM SCHAALII
#> 6          TETRATHIOBACTER KASHMIRENSIS      ADVENELLA KASHMIRENSIS
df$species <- lookup_recode(df$spec,'species')
df[,c('spec','species','date')]
#>                            spec                     species       date
#> 3          CHRYSEOMONAS LUTEOLA         PSEUDOMONAS LUTEOLA 2021-02-27
#> 10             ESCHERICHIA COLI            ESCHERICHIA COLI 2021-04-12
#> 9         STREPTOCOCCUS GROUP D       ENTEROCOCCUS FAECALIS 2021-04-16
#> 2  TETRATHIOBACTER KASHMIRENSIS      ADVENELLA KASHMIRENSIS 2021-05-01
#> 7         ANAEROCOCCUS PREVOTII PEPTOSTREPTOCOCCUS PREVOTII 2021-05-16
#> 11                   SARS-COV-2                  SARS-COV-2 2021-05-24
#> 5              STOMATOCOCCUS SP                   ROTHIA SP 2021-08-04
#> 1       PROPIONIBACTERIUM ACNES         CUTIBACTERIUM ACNES 2021-08-07
#> 6       RHODOCOCCUS BRONCHIALIS        GORDONIA BRONCHIALIS 2021-08-21
#> 12               CANDIDA AUREUS              CANDIDA AUREUS 2021-08-25
#> 8             EGGERTHELLA LENTA           EGGERTHELLA LENTA 2021-08-29
#> 4      PSEUDOMONAS PAUCIMOBILIS   SPHINGOMONAS PAUCIMOBILIS 2022-02-07

head(specimen_type_grouping)
#>      specimen_type specimen_group
#> 1            BLOOD          Blood
#> 2       CORD BLOOD          Blood
#> 3            JOINT Bones & Joints
#> 4   SYNOVIAL FLUID Bones & Joints
#> 5             BONE Bones & Joints
#> 6 JOINT PROSTHESIS Bones & Joints
df$grp <- lookup_recode(df$type,'specimen')
df[,c('species','type','grp','date')]
#>                        species                    type            grp
#> 3          PSEUDOMONAS LUTEOLA                  VAGINA        Genital
#> 10            ESCHERICHIA COLI              BIOPSY-NOS  Swabs-General
#> 9        ENTEROCOCCUS FAECALIS                  CERVIX        Genital
#> 2       ADVENELLA KASHMIRENSIS                  SALIVA  URT/Mouth/Ear
#> 7  PEPTOSTREPTOCOCCUS PREVOTII                  THROAT  URT/Mouth/Ear
#> 11                  SARS-COV-2        JOINT PROSTHESIS Bones & Joints
#> 5                    ROTHIA SP ENDOTRACHEAL SECRETIONS  URT/Mouth/Ear
#> 1          CUTIBACTERIUM ACNES                PERINEUM        Tissues
#> 6         GORDONIA BRONCHIALIS   ENDOTRACHEAL ASPIRATE  URT/Mouth/Ear
#> 12              CANDIDA AUREUS       HEART/HEART VALVE        Tissues
#> 8            EGGERTHELLA LENTA                    BURN  Swabs-General
#> 4    SPHINGOMONAS PAUCIMOBILIS        WOUND (SURGICAL)    Wound Swabs
#>          date
#> 3  2021-02-27
#> 10 2021-04-12
#> 9  2021-04-16
#> 2  2021-05-01
#> 7  2021-05-16
#> 11 2021-05-24
#> 5  2021-08-04
#> 1  2021-08-07
#> 6  2021-08-21
#> 12 2021-08-25
#> 8  2021-08-29
#> 4  2022-02-07

# for a tidyverse use
# df %>% mutate(spec=lookup_recode(spec,'species))

# manual input of your own lookup
# .import=list(new,old)
lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
#> [[1]]
#> [1] "ACHROMOBACTER DENITRIFICANS"
#>