Lookup table switch handler

A function to call an epidm lookup table and recode where we are aware of a new value.

Built in are the organism re-classifications and specimen_type groupings and a manual mode.

lookup_recode(src, type = c("species", "specimen", "manual"), .import = NULL)

Arguments

src

a character, vector or column containing the value(s) to be referenced

type

a character value to denote the lookup table used

.import

a list in the order list(new,old) containing the values for another lookup table existing in the environment

Value

a list object of the recoded field

Examples

df <- data.frame(
  spec = c(
    sample(grep(")",
                respeciate_organism$previous_organism_name,
                value=TRUE,
                invert = TRUE),
           9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  type = sample(specimen_type_grouping$specimen_type,12),
  date = sample(seq.Date(from = Sys.Date()-365,
                         to = Sys.Date(),
                         by = "day"),12)
)
df <- df[order(df$date),]

# show the data before the changes
df
#>                           spec                                 type       date
#> 9    ALCALIGENES DENITRIFICANS                         URINE/KIDNEY 2021-02-16
#> 4      FLAVOBACTERIUM ODORATUM                ENDOTRACHEAL ASPIRATE 2021-03-14
#> 2       ALCALIGENES PIECHAUDII                            DRAINSITE 2021-05-28
#> 8  STOMATOCOCCUS MUCILAGINOSUS                                 BURN 2021-06-03
#> 1      RHODOCOCCUS BRONCHIALIS                                  LVS 2021-06-05
#> 10            ESCHERICHIA COLI                                STOOL 2021-07-06
#> 3    PEPTOSTREPTOCOCCUS MAGNUS                          CHEST DRAIN 2021-08-07
#> 5      PLURALIBACTER GERGOVIAE                              VENFLON 2021-09-09
#> 6   STREPTOCOCCUS GROUP D STEM                         GALL BLADDER 2021-09-10
#> 7          PANTOEA AGGLOMERANS                                  NPA 2021-11-20
#> 11                  SARS-COV-2                                 LUNG 2022-01-08
#> 12              CANDIDA AUREUS FAECES/LOWER GASTRO-INTESTINAL TRACT 2022-01-13

# check the lookup tables
# observe the changes
head(respeciate_organism[1:2])
#>                  previous_organism_name       organism_species_name
#> 1             ALCALIGENES DENITRIFICANS ACHROMOBACTER DENITRIFICANS
#> 2                ALCALIGENES PIECHAUDII    ACHROMOBACTER PIECHAUDII
#> 3              ALCALIGENES XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 4 ALCALIGENES XYLOSOXIDANS XYLOSOXIDANS  ACHROMOBACTER XYLOSOXIDANS
#> 5                ACTINOBACULUM SCHAALII       ACTINOTIGNUM SCHAALII
#> 6          TETRATHIOBACTER KASHMIRENSIS      ADVENELLA KASHMIRENSIS
df$species <- lookup_recode(df$spec,'species')
df[,c('spec','species','date')]
#>                           spec                     species       date
#> 9    ALCALIGENES DENITRIFICANS ACHROMOBACTER DENITRIFICANS 2021-02-16
#> 4      FLAVOBACTERIUM ODORATUM           MYROIDES ODORATUS 2021-03-14
#> 2       ALCALIGENES PIECHAUDII    ACHROMOBACTER PIECHAUDII 2021-05-28
#> 8  STOMATOCOCCUS MUCILAGINOSUS         ROTHIA MUCILAGINOSA 2021-06-03
#> 1      RHODOCOCCUS BRONCHIALIS        GORDONIA BRONCHIALIS 2021-06-05
#> 10            ESCHERICHIA COLI            ESCHERICHIA COLI 2021-07-06
#> 3    PEPTOSTREPTOCOCCUS MAGNUS            FINEGOLDIA MAGNA 2021-08-07
#> 5      PLURALIBACTER GERGOVIAE      ENTEROBACTER GERGOVIAE 2021-09-09
#> 6   STREPTOCOCCUS GROUP D STEM       ENTEROCOCCUS FAECALIS 2021-09-10
#> 7          PANTOEA AGGLOMERANS         PANTOEA AGGLOMERANS 2021-11-20
#> 11                  SARS-COV-2                  SARS-COV-2 2022-01-08
#> 12              CANDIDA AUREUS              CANDIDA AUREUS 2022-01-13

head(specimen_type_grouping)
#>      specimen_type specimen_group
#> 1            BLOOD          Blood
#> 2       CORD BLOOD          Blood
#> 3            JOINT Bones & Joints
#> 4   SYNOVIAL FLUID Bones & Joints
#> 5             BONE Bones & Joints
#> 6 JOINT PROSTHESIS Bones & Joints
df$grp <- lookup_recode(df$type,'specimen')
df[,c('species','type','grp','date')]
#>                        species                                 type
#> 9  ACHROMOBACTER DENITRIFICANS                         URINE/KIDNEY
#> 4            MYROIDES ODORATUS                ENDOTRACHEAL ASPIRATE
#> 2     ACHROMOBACTER PIECHAUDII                            DRAINSITE
#> 8          ROTHIA MUCILAGINOSA                                 BURN
#> 1         GORDONIA BRONCHIALIS                                  LVS
#> 10            ESCHERICHIA COLI                                STOOL
#> 3             FINEGOLDIA MAGNA                          CHEST DRAIN
#> 5       ENTEROBACTER GERGOVIAE                              VENFLON
#> 6        ENTEROCOCCUS FAECALIS                         GALL BLADDER
#> 7          PANTOEA AGGLOMERANS                                  NPA
#> 11                  SARS-COV-2                                 LUNG
#> 12              CANDIDA AUREUS FAECES/LOWER GASTRO-INTESTINAL TRACT
#>                        grp       date
#> 9             Urine/Kidney 2021-02-16
#> 4            URT/Mouth/Ear 2021-03-14
#> 2                   Fluids 2021-05-28
#> 8            Swabs-General 2021-06-03
#> 1                  Genital 2021-06-05
#> 10       Faeces & Lowergut 2021-07-06
#> 3                   Fluids 2021-08-07
#> 5             Tips & Lines 2021-09-09
#> 6        Faeces & Lowergut 2021-09-10
#> 7            URT/Mouth/Ear 2021-11-20
#> 11 Lower Respiratory Tract 2022-01-08
#> 12       Faeces & Lowergut 2022-01-13

# for a tidyverse use
# df %>% mutate(spec=lookup_recode(spec,'species))

# manual input of your own lookup
# .import=list(new,old)
lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )
#> [[1]]
#> [1] "ACHROMOBACTER DENITRIFICANS"
#>