Lookup table switch handler

A function to call an epidm lookup table and recode where we are aware of a new value.

Built in are the organism re-classifications and specimen_type groupings and a manual mode.

lookup_recode(src, type = c("species", "specimen", "manual"), .import = NULL)

Arguments

src
a character, vector or column containing the value(s) to be referenced

type
a character value to denote the lookup table used

.import
a list in the order list(new,old) containing the values for another lookup table existing in the environment

Value

a list object of the recoded field

Examples

df <- data.frame(
  sp = c(
    sample(respeciate_organism$previous_organism_name,9),
    "ESCHERICHIA COLI","SARS-COV-2","CANDIDA AUREUS"),
  ty = sample(specimen_type_grouping$specimen_type,12),
  dt = sample(seq.Date(from = Sys.Date()-365,
                       to = Sys.Date(),
                       by = "day"),12)
)
df <- df[order(df$dt),]

df$species <- lookup_recode(df$sp,'species')
df$grp <- lookup_recode(df$ty,'specimen')
df

#>                                                sp                ty         dt
#> 8                      STREPTOCOCCUS GROUP D STEM         BRONCHIAL 2020-04-26
#> 7                     STOMATOCOCCUS MUCILAGINOSUS              BURN 2020-05-07
#> 9           ALCALIGENES XYLOSOXIDANS XYLOSOXIDANS             WOUND 2020-06-03
#> 11                                     SARS-COV-2             PENIS 2020-08-22
#> 3      ANAEROCOCCUS (PEPTOSTREPTOCOCCUS) PREVOTII   IV CATHETER TIP 2020-09-03
#> 4                                 VIBRIO HOLLISAE ARTERIAL LINE TIP 2020-09-04
#> 6                           ANAEROCOCCUS PREVOTII        SKIN ULCER 2020-10-25
#> 5                                  HANIA PARALVEI    DENTAL ABSCESS 2020-10-28
#> 2                              EUBACTERIUM LENTUM  PERITONEAL FLUID 2020-11-20
#> 1  ENTEROBACTER AGGLOMERANS (PANTOEA AGGLOMERANS)              BONE 2020-12-10
#> 10                               ESCHERICHIA COLI           TRACHEA 2020-12-15
#> 12                                 CANDIDA AUREUS        BIOPSY-NOS 2020-12-16
#>                        species                     grp
#> 8        ENTEROCOCCUS FAECALIS Lower Respiratory Tract
#> 7          ROTHIA MUCILAGINOSA           Swabs-General
#> 9   ACHROMOBACTER XYLOSOXIDANS             Wound Swabs
#> 11                  SARS-COV-2                 Genital
#> 3        ANAEROCOCCUS PREVOTII            Tips & Lines
#> 4           GRIMONTIA HOLLISAE            Tips & Lines
#> 6  PEPTOSTREPTOCOCCUS PREVOTII           Swabs-General
#> 5                 HAVNIA ALVEI           URT/Mouth/Ear
#> 2            EGGERTHELLA LENTA                  Fluids
#> 1          PANTOEA AGGLOMERANS          Bones & Joints
#> 10            ESCHERICHIA COLI Lower Respiratory Tract
#> 12              CANDIDA AUREUS           Swabs-General

# for a tidyverse use
# df %>% mutate(sp=lookup_recode(sp,'species))

lookup_recode(
  "ALCALIGENES DENITRIFICANS",
  type = 'manual',
  .import=list(respeciate_organism$organism_species_name,
               respeciate_organism$previous_organism_name)
  )

#> [[1]]
#> [1] "ACHROMOBACTER DENITRIFICANS"
#>