Respeciate unspecified samples

Some samples within SGSS are submitted by laboratories as "GENUS SP" or "GENUS UNNAMED". However, they may also have a fully identified sample taken from the same site within a recent time period. This function captures species from another sample within X-days of an unspeciated isolate.

respeciate_generic(x, group_vars, species, date, window = c(0:Inf))

Arguments

x
a data.frame or data.table object

group_vars
the minimum grouping set of variables for like samples in a character vector; suggest c('patient_id','specimen_type','genus')

species
a character containing the column with the organism species name

date
a character containing the column with the specimen/sample date

window
an integer representing the number of days for which you will allow a sample to be respeciated

Value

a data.table with a recharacterised specimen column

Examples

df <- data.frame(
ptid = c(round(runif(25,1,5))),
spec = sample(c("KLEBSIELLA SP",
                "KLEBSIELLA UNNAMED",
                "KLEBSIELLA PNEUMONIAE",
                "KLEBEIELLA OXYTOCA"),
              25,replace = T),
type = "BLOOD",
date = sample(seq.Date(Sys.Date()-21,Sys.Date(),"day"),25,replace = T)
)

respeciate_generic(x=df,
                   group_vars=c('ptid','type'),
                   species='spec',
                   date='date',
                   window = 14)[]

#> Error in respeciate_generic(x = df, group_vars = c("ptid", "type"), species = "spec",     date = "date", window = 14): object '.forceCopy' not found