Respeciate unspecified samples

Some samples within SGSS are submitted by laboratories as "GENUS SP" or "GENUS UNNAMED". However, they may also have a fully identified sample taken from the same site within a recent time period. This function captures species_col from another sample within X-days of an unspeciated isolate.

respeciate_generic(
  x,
  group_vars,
  species_col,
  date_col,
  window = c(0:Inf),
  .forceCopy = FALSE
)

Arguments

x

a data.frame or data.table object

group_vars

the minimum grouping set of variables for like samples in a character vector; suggest c('patient_id','specimen_type','genus')

species_col

a character containing the column with the organism species_col name

date_col

a character containing the column with the specimen/sample date_col

window

an integer representing the number of days for which you will allow a sample to be respeciated

.forceCopy

default FALSE; TRUE will force data.table to take a copy instead of editing the data without reference

Value

a data.table with a recharacterised species_col column

Examples

df <- data.frame(
ptid = c(round(runif(25,1,5))),
spec = sample(c("KLEBSIELLA SP",
                "KLEBSIELLA UNNAMED",
                "KLEBSIELLA PNEUMONIAE",
                "KLEBEIELLA OXYTOCA"),
              25,replace = TRUE),
type = "BLOOD",
specdate = sample(seq.Date(Sys.Date()-21,Sys.Date(),"day"),25,replace = TRUE)
)

respeciate_generic(x=df,
                   group_vars=c('ptid','type'),
                   species_col='spec',
                   date_col='specdate',
                   window = 14)[]
#> 14-day round 1: 7 SP or UNNAMMED isolates respeciated
#> 14-day round 2: 2 SP or UNNAMMED isolates respeciated
#> 14-day round 3: 0 SP or UNNAMMED isolates respeciated
#>     ptid                  spec  type   specdate
#>  1:    1 KLEBSIELLA PNEUMONIAE BLOOD 2022-01-31
#>  2:    1 KLEBSIELLA PNEUMONIAE BLOOD 2022-01-31
#>  3:    1 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-07
#>  4:    1 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-09
#>  5:    2    KLEBEIELLA OXYTOCA BLOOD 2022-02-05
#>  6:    2    KLEBEIELLA OXYTOCA BLOOD 2022-02-05
#>  7:    2 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-13
#>  8:    2 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-13
#>  9:    2 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-14
#> 10:    3    KLEBEIELLA OXYTOCA BLOOD 2022-01-24
#> 11:    3    KLEBEIELLA OXYTOCA BLOOD 2022-01-28
#> 12:    3    KLEBEIELLA OXYTOCA BLOOD 2022-01-29
#> 13:    3    KLEBEIELLA OXYTOCA BLOOD 2022-02-02
#> 14:    3 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-10
#> 15:    3    KLEBEIELLA OXYTOCA BLOOD 2022-02-10
#> 16:    3    KLEBEIELLA OXYTOCA BLOOD 2022-02-10
#> 17:    4 KLEBSIELLA PNEUMONIAE BLOOD 2022-01-27
#> 18:    4    KLEBEIELLA OXYTOCA BLOOD 2022-01-28
#> 19:    4    KLEBEIELLA OXYTOCA BLOOD 2022-01-31
#> 20:    4    KLEBEIELLA OXYTOCA BLOOD 2022-02-01
#> 21:    4 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-03
#> 22:    4 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-10
#> 23:    5    KLEBEIELLA OXYTOCA BLOOD 2022-02-08
#> 24:    5    KLEBEIELLA OXYTOCA BLOOD 2022-02-09
#> 25:    5 KLEBSIELLA PNEUMONIAE BLOOD 2022-02-14
#>     ptid                  spec  type   specdate