Patient ID record grouping

Groups patient records from multiple isolates with a single integer patientID by grouping patient identifiers.

Grouping is based on five stages:

matching nhs number and date of birth

Hospital number & Date of Birth

NHS number & Hospital Number

Date of Birth & Surname IF nhs unknown

Sex & Date of Birth & Fuzzy Name

uk_patient_id(
  x,
  nhs_number,
  hospital_number,
  date_of_birth,
  sex,
  forename = "NONAME",
  surname = "NONAME",
  .forceCopy = FALSE
)

Arguments

x
a data.frame or data.table containing the cleaned line list

nhs_number
a column as a character containing the patient NHS numbers

hospital_number
a column as a character containing the patient Hospital numbers

date_of_birth
a column as a date variable containing the patient date of birth in date format

sex
a column as a character containing the patient sex

forename
a column as a character containing the patient forename; leave as NONAME if unavailable

surname
a column as a character containing the patient surname; leave as NONAME if unavailable

.forceCopy
default FALSE; TRUE will force data.table to take a copy instead of editing the data without reference

Value

patientID grouping variable addedd to the data.frame

A dataframe with two new variables: id a unique patient id, and n_in_id an integer variable with the number of rows in the id

Examples

id_test <- data.frame(
  nhs_n = c(
    9434765919,9434765919,9434765919,NA,NA,
    3367170666,5185293519,5185293519,5185293519,8082318562,NA,NA,NA
  ),
  hosp_n = c(
    '13','13','13','UNKNOWN','13','13','13','31','31','96','96',NA,'96'),
  sex = c('M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'M', 'F', 'U', 'U', 'F'),
  dateofbirth = as.Date(
    c(
      '1988-10-06','1988-06-10','1900-01-01','1988-10-06','1988-10-06',
      '1988-10-06','1988-10-06','1988-10-06','1988-10-06','2020-01-28',
      '2020-01-28','2020-01-28','2020-01-28'
    )
  ),
  firstname = c(
    'Danger','Danger','Denger','Danger','Danger','DANGER','Danger',
    'Danger','Danger','Crazy','Crazy','Krazy','C'
  ),
  lastname = c(
    'Mouse','Mause','Mouse','Moose','Moose','Mouse','MOUSe',
    'Mouse','Mouse','Frog','FROG','Frug','Frog'
  )
)
uk_patient_id(x = id_test,
              nhs_number = 'nhs_n',
              hospital_number = 'hosp_n',
              forename = 'firstname',
              surname = 'lastname',
              sex = 'sex',
              date_of_birth = 'dateofbirth')[]

#>     id      nhs_n  hosp_n sex dateofbirth firstname lastname
#>  1:  1 9434765919      13   M  1988-10-06    DANGER    MOUSE
#>  2:  1 9434765919      13   M  1988-06-10    DANGER    MAUSE
#>  3:  1 9434765919      13   M  1900-01-01    DENGER    MOUSE
#>  4:  1         NA UNKNOWN   M  1988-10-06    DANGER    MOOSE
#>  5:  1         NA      13   M  1988-10-06    DANGER    MOOSE
#>  6:  1 3367170666      13   M  1988-10-06    DANGER    MOUSE
#>  7:  1 5185293519      13   M  1988-10-06    DANGER    MOUSE
#>  8:  7 5185293519      31   M  1988-10-06    DANGER    MOUSE
#>  9:  7 5185293519      31   M  1988-10-06    DANGER    MOUSE
#> 10: 10 8082318562      96   F  2020-01-28     CRAZY     FROG
#> 11: 10         NA      96   U  2020-01-28     CRAZY     FROG
#> 12: 12         NA    <NA>   U  2020-01-28     KRAZY     FRUG
#> 13: 10         NA      96   F  2020-01-28         C     FROG